Separate TTLBucket and TTLBucketList to standalone files
